requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2>=2.7.5,<2.8.0
requests==2.25.1

flake8>=3.6.0,<3.7.0